Refactor channel classes to have separate pub/req channels
