+++ salt/transport/zeromq.py
@@ -1,0 +1,124 @@
+'''
+Zeromq transport classes
+'''
+
+import os
+import threading
+
+# Import Salt Libs
+import salt.payload
+import salt.auth
+import salt.crypt
+import salt.utils
+import logging
+from collections import defaultdict
+
+from salt.utils import kinds
+from salt.transport.channel import ReqChannel
+
+log = logging.getLogger(__name__)
+
+class ZeroMQReqChannel(ReqChannel):
+    '''
+    Encapsulate sending routines to ZeroMQ.
+
+    ZMQ Channels default to 'crypt=aes'
+    '''
+    # the sreq is the zmq connection, since those are relatively expensive to
+    # set up, we are going to reuse them as much as possible.
+    sreq_cache = defaultdict(dict)
+
+    @property
+    def sreq_key(self):
+        '''
+        Return a tuple which uniquely defines this channel (for caching)
+        '''
+        return (self.master_uri,                  # which master you want to talk to
+                os.getpid(),                      # per process
+                threading.current_thread().name,  # per per-thread
+                )
+
+    @property
+    def sreq(self):
+        # When using threading, like on Windows, don't cache.
+        # The following block prevents thread leaks.
+        if not self.opts.get('multiprocessing'):
+            return salt.payload.SREQ(self.master_uri)
+
+        key = self.sreq_key
+
+        if not self.opts['cache_sreqs']:
+            return salt.payload.SREQ(self.master_uri)
+        else:
+            if key not in ZeroMQReqChannel.sreq_cache:
+                master_type = self.opts.get('master_type', None)
+                if master_type == 'failover':
+                    # remove all cached sreqs to the old master to prevent
+                    # zeromq from reconnecting to old masters automagically
+                    for check_key in self.sreq_cache.keys():
+                        if self.opts['master_uri'] != check_key[0]:
+                            del self.sreq_cache[check_key]
+                            log.debug('Removed obsolete sreq-object from '
+                                      'sreq_cache for master {0}'.format(check_key[0]))
+
+                ZeroMQReqChannel.sreq_cache[key] = salt.payload.SREQ(self.master_uri)
+
+            return ZeroMQReqChannel.sreq_cache[key]
+
+    def __init__(self, opts, **kwargs):
+        self.opts = opts
+        self.ttype = 'zeromq'
+
+        # crypt defaults to 'aes'
+        self.crypt = kwargs.get('crypt', 'aes')
+
+        if 'master_uri' in kwargs:
+            self.master_uri = kwargs['master_uri']
+        else:
+            self.master_uri = opts['master_uri']
+
+        if self.crypt != 'clear':
+            # we don't need to worry about auth as a kwarg, since its a singleton
+            self.auth = salt.crypt.SAuth(self.opts)
+
+    def crypted_transfer_decode_dictentry(self, load, dictkey=None, tries=3, timeout=60):
+        ret = self.sreq.send('aes', self.auth.crypticle.dumps(load), tries, timeout)
+        key = self.auth.get_keys()
+        aes = key.private_decrypt(ret['key'], 4)
+        pcrypt = salt.crypt.Crypticle(self.opts, aes)
+        return pcrypt.loads(ret[dictkey])
+
+    def _crypted_transfer(self, load, tries=3, timeout=60):
+        '''
+        In case of authentication errors, try to renegotiate authentication
+        and retry the method.
+        Indeed, we can fail too early in case of a master restart during a
+        minion state execution call
+        '''
+        def _do_transfer():
+            data = self.sreq.send(
+                self.crypt,
+                self.auth.crypticle.dumps(load),
+                tries,
+                timeout)
+            # we may not have always data
+            # as for example for saltcall ret submission, this is a blind
+            # communication, we do not subscribe to return events, we just
+            # upload the results to the master
+            if data:
+                data = self.auth.crypticle.loads(data)
+            return data
+        try:
+            return _do_transfer()
+        except salt.crypt.AuthenticationError:
+            self.auth.authenticate()
+            return _do_transfer()
+
+    def _uncrypted_transfer(self, load, tries=3, timeout=60):
+        return self.sreq.send(self.crypt, load, tries, timeout)
+
+    def send(self, load, tries=3, timeout=60):
+        if self.crypt == 'clear':  # for sign-in requests
+            return self._uncrypted_transfer(load, tries, timeout)
+        else:  # for just about everything else
+            return self._crypted_transfer(load, tries, timeout)
